requirements.txt
# LangChain Core
langchain
langchain-core

# Groq Integration
langchain-groq
groq

# Anthropic Integration
#langchain-anthropic

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

sentence-transformers

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# streamlit # UI interface
